requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==4.0.1
blinker==1.9.0
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.119.1
Flask==3.1.2
folium==0.20.0
geographiclib==2.1
geopy==2.4.1
gpxpy==1.6.2
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
numpy==2.3.4
orjson==3.11.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.15.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xyzservices==2025.10.0
